Use confirmation for deleting

+++ app/controllers/tasks/task/show.js
@@ -3,9 +3,11 @@ import Ember from 'ember';
 export default Ember.Controller.extend({
   actions: {
     destroyTask() {
-      this.get('model').destroyRecord().then(() => {
-        this.transitionToRoute('tasks');
-      });
+      if (confirm('Are you sure?')) {
+        this.get('model').destroyRecord().then(() => {
+          this.transitionToRoute('tasks');
+        });
+      }
     }
   }
 });
